Moved todo mutations to mutations directory. removed starter data
